% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twalk_engine.R
\name{executar_movimento}
\alias{executar_movimento}
\title{Executa um único movimento (passo) do algoritmo t-walk}
\usage{
executar_movimento(
  dimensao,
  funcao_objetivo,
  funcao_suporte,
  x,
  valor_U,
  x_prime,
  valor_Up,
  parametro_at = 6,
  parametro_aw = 1.5,
  prob_phi = min(dimensao, 4)/dimensao,
  freq_1 = 0.4918,
  freq_2 = 0.9836,
  freq_3 = 0.9918,
  ...
)
}
\arguments{
\item{...}{Argumentos passados para `funcao_objetivo` e `funcao_suporte`.}
}
\value{
Uma lista contendo a proposta e a probabilidade de aceitação.
}
\description{
Seleciona um dos quatro kernels, gera uma proposta e calcula
a probabilidade de aceitação de Metropolis-Hastings.
}
\keyword{internal}
