% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twalk_engine.R
\name{kernel_blow}
\alias{kernel_blow}
\title{Gera proposta para o kernel Blow}
\usage{
kernel_blow(dimensao, prob_phi, x, x_prime)
}
\arguments{
\item{dimensao}{Dimensão do espaço de parâmetros.}

\item{prob_phi}{Probabilidade de atualizar cada coordenada.}

\item{x}{Ponto atual `x`.}

\item{x_prime}{Ponto atual `x'`.}
}
\value{
Uma lista com `proposta`, `n_phi` e `phi`.
}
\description{
Gera proposta para o kernel Blow
}
\keyword{internal}
