% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twalk_engine.R
\name{kernel_traverse}
\alias{kernel_traverse}
\title{Gera proposta para o kernel Traverse}
\usage{
kernel_traverse(dimensao, prob_phi, x, x_prime, beta)
}
\arguments{
\item{dimensao}{Dimensão do espaço de parâmetros.}

\item{prob_phi}{Probabilidade de atualizar cada coordenada.}

\item{x}{Ponto atual `x`.}

\item{x_prime}{Ponto atual `x'`.}

\item{beta}{Parâmetro de passo beta.}
}
\value{
Uma lista com `proposta` e `n_phi` (número de coordenadas movidas).
}
\description{
Gera proposta para o kernel Traverse
}
\keyword{internal}
