% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twalk_engine.R
\name{kernel_walk}
\alias{kernel_walk}
\title{Gera proposta para o kernel Walk}
\usage{
kernel_walk(dimensao, prob_phi, parametro_aw, x, x_prime)
}
\arguments{
\item{dimensao}{Dimensão do espaço de parâmetros.}

\item{prob_phi}{Probabilidade de atualizar cada coordenada.}

\item{parametro_aw}{Parâmetro de escala `a_w` (padrão 1.5).}

\item{x}{Ponto atual `x`.}

\item{x_prime}{Ponto atual `x'`.}
}
\value{
Uma lista com `proposta` e `n_phi`.
}
\description{
Gera proposta para o kernel Walk
}
\keyword{internal}
