% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twalk_engine.R
\name{log_densidade_hop}
\alias{log_densidade_hop}
\title{Calcula o log da densidade da proposta para o kernel Hop}
\usage{
log_densidade_hop(n_phi, phi, h, x, x_prime)
}
\arguments{
\item{n_phi}{Número de coordenadas movidas.}

\item{phi}{Vetor booleano indicando as coordenadas movidas.}

\item{h}{Ponto proposto.}

\item{x}{Ponto de referência.}

\item{x_prime}{Ponto de referência.}
}
\value{
O valor de -log(g(h)).
}
\description{
Calcula o log da densidade da proposta para o kernel Hop
}
\keyword{internal}
