% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twalk_engine.R
\name{simular_beta}
\alias{simular_beta}
\title{Simula o parâmetro beta para o kernel Traverse}
\usage{
simular_beta(at)
}
\arguments{
\item{at}{Parâmetro de forma `a_t` (padrão 6).}
}
\value{
Um valor numérico para beta.
}
\description{
Amostra beta da distribuição proposta na Seção 2.2 de Christen & Fox (2010).
}
\keyword{internal}
